Fix app source PR review and CI coverage export
