+++ api/tests/unit/services/mcp_server/test_tool_implementations.py
@@ -166,7 +166,8 @@ class _CreateAppDb:
         self.flushed = True
 
     async def commit(self):
-        pass
+        # Intentionally no-op: the fake only records whether create_app reaches commit.
+        return None
 
 
 class TestAppToolImpl:
